+++ parser.test.ts
@@ -68,6 +68,7 @@ type _Tests = [
   Expect<Equal<ParseSelector<'div[data-key*=value]'>, HTMLDivElement>>,
   Expect<Equal<ParseSelector<'div[data-key="value"i]'>, HTMLDivElement>>,
   Expect<Equal<ParseSelector<'div[data-key="value"s]'>, HTMLDivElement>>,
+  Expect<Equal<ParseSelector<'div[]'>, Element>>,
   Expect<
     Equal<ParseSelector<'link[type="application/atom+xml"]'>, HTMLLinkElement>
   >,